move point-of-service related models to separate file and rename store.model to store-finder.model
