+++ projects/core/src/store-finder/connectors/store-finder.adapter.ts
@@ -1,11 +1,11 @@
 import { StoreFinderSearchConfig } from '../model/search-config';
 import { Observable } from 'rxjs';
+import { PointOfService } from '../../model/point-of-service.model';
+import { GeoPoint } from '../../model/misc.model';
 import {
   StoreFinderSearchPage,
-  PointOfService,
   StoreCount,
-} from '../../model/store.model';
-import { GeoPoint } from '../../model/misc.model';
+} from '../../model/store-finder.model';
 
 export abstract class StoreFinderAdapter {
   abstract search(
