+++ projects/core/src/store-finder/model/store-entities.ts
@@ -1,4 +1,4 @@
-import { PointOfService } from '../../model/store.model';
+import { PointOfService } from '../../model/point-of-service.model';
 
 export interface StoreEntities {
   pointOfServices?: Array<PointOfService>;
